fix problems mentioned in comments of PR
