--- src/app/commons/ProblemEditorPage/ProblemEditorPage.css
@@ -1,113 +1,0 @@
-.problem-editor .tag-input {
-    width: 250px;
-    height: 33px;
-    border: 1px solid #E4E1E1;
-    padding: 5px 10px;
-}
-
-.problem-editor .tag-button {
-    width: 33px;
-    height: 33px;
-
-    /* Add Btn */
-    /* Auto Layout */
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 4px;
-
-    background: #F9F9F9;
-    border: 1px solid #E5E5E5;
-    box-sizing: border-box;
-}
-
-.problem-editor .tag-item {
-    color: #1273BA;
-    padding-left: 10px;
-    padding-top: 5px;
-    padding-bottom: 5px;
-    margin-top: 10px;
-    margin-bottom: 10px;
-    width: 284px;
-    height: 33px;
-    border: 1px solid #E4E1E1;
-    box-sizing: border-box;
-
-    font-family: Poppins, sans-serif;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 15px;
-    line-height: 22px;
-
-    display: flex;
-    align-items: center;
-}
-
-.problem-editor .delete-icon {
-    position: absolute;
-    width: 10px;
-    height: 10px;
-    float: right;
-    left: 400px;
-}
-
-.problem-editor .input-box {
-    /* Input Box - small */
-    /* Auto Layout */
-    flex-direction: row;
-    align-items: center;
-    padding: 2px 5px 2px 10px;
-
-    position: static;
-    width: 75px;
-    height: 24px;
-    left: 51px;
-    top: 0px;
-
-    background: #F3F3F3;
-    border: 1px solid #C4C4C4;
-    box-sizing: border-box;
-    border-radius: 5px;
-
-    /* Inside Auto Layout */
-    flex: none;
-    order: 1;
-    flex-grow: 0;
-    margin: 0px 7px;
-}
-
-.problem-editor li {
-    color: #1273BA;
-    /* Inside Auto Layout */
-
-    flex: none;
-    order: 0;
-    flex-grow: 0;
-    margin: 0px 5px;
-    padding-bottom: 10px;
-}
-
-.problem-editor li span {
-    position: relative;
-    left: -10px;
-}
-
-/* to hide original upload button */
-.problem-editor .original-upload-button {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0,0,0,0);
-    border: 0;
-}
-
-/* TODO: apply in the UploadFileButton */
-/*.problem-editor .provided-code-button:hover {*/
-/*    color: white;*/
-/*    font-weight: bold;*/
-/*    background: rgb(31, 152, 178);*/
-/*}*/
