+++ src/app/commons/TextInputForm/TextInputForm.css
@@ -1,0 +1,23 @@
+.problem-editor .tag-input {
+    width: 250px;
+    height: 33px;
+    border: 1px solid #E4E1E1;
+    padding: 5px 10px;
+}
+
+.problem-editor .tag-button {
+    width: 33px;
+    height: 33px;
+
+    /* Add Btn */
+    /* Auto Layout */
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    padding: 4px;
+
+    background: #F9F9F9;
+    border: 1px solid #E5E5E5;
+    box-sizing: border-box;
+}
