--- src/app/problem/deleteProblemButton/DeleteProblemButton.js
@@ -1,11 +1,0 @@
-import {EditorButton} from "../../commons/ProblemEditorPage/ProblemEditorPage";
-
-function SaveChangeButton () {
-    return (
-        <EditorButton
-            text={"Delete Problem"} buttonColor={"#FFFFFF"} fontColor={"#A2A3B1"}
-        />
-    )
-}
-
-export default SaveChangeButton;
