+++ src/app/problem/edit/CompilationScript.css
@@ -1,0 +1,37 @@
+.problem-editor .compile-script-button {
+    /* Button */
+    position: static;
+    width: 89px;
+    height: 18px;
+    left: 6px;
+    top: 2px;
+
+    font-family: Poppins, sans-serif;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 12px;
+    line-height: 18px;
+
+    display: flex;
+    align-items: center;
+
+    color: #3EBDD9;
+
+    /* Inside Auto Layout */
+    flex: none;
+    order: 0;
+    flex-grow: 0;
+    margin: 0px 5px;
+
+    background: #FFFFFF;
+    border: 1px solid #3EBDD9;
+    box-sizing: border-box;
+    border-radius: 10px;
+}
+
+.problem-editor .compile-script-text-area {
+    background: #F9F9F9;
+    border: 1px solid #C4C4C4;
+    box-sizing: border-box;
+    border-radius: 5px;
+}
