+++ src/app/problem/edit/TagList.js
@@ -2,7 +2,8 @@ import React from 'react';
 import '../ProblemEditor.css';
 import {useTags} from "../../usecases/TagUseCase";
 import {SubtitleLine} from "../../commons/titles/TitleLine";
-import {InputForm, Items} from '../../commons/ProblemEditorPage/ProblemEditorPage';
+import {TextInputForm} from "../../commons/TextInputForm/TextInputForm";
+import {TextInputItems} from "./TextInputItems";
 
 function TagList() {
     const {tags, addTags, removeTag} = useTags();
@@ -10,8 +11,8 @@ function TagList() {
     return (
         <div>
             <SubtitleLine title={"Tags"} />
-            <InputForm placeholder={"Add New Tags"} onSubmit={addTags}  />
-            <Items items={tags} removeItems={removeTag} />
+            <TextInputForm placeholder={"Add New Tags"} onSubmit={addTags}  />
+            <TextInputItems items={tags} removeItems={removeTag} />
         </div>
     )
 }
