+++ src/app/problem/edit/TextInputItems.css
@@ -1,0 +1,29 @@
+.problem-editor .tag-item {
+    color: #1273BA;
+    padding-left: 10px;
+    padding-top: 5px;
+    padding-bottom: 5px;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    width: 284px;
+    height: 33px;
+    border: 1px solid #E4E1E1;
+    box-sizing: border-box;
+
+    font-family: Poppins, sans-serif;
+    font-style: normal;
+    font-weight: normal;
+    font-size: 15px;
+    line-height: 22px;
+
+    display: flex;
+    align-items: center;
+}
+
+.problem-editor .delete-icon {
+    position: absolute;
+    width: 10px;
+    height: 10px;
+    float: right;
+    left: 400px;
+}
