--- src/app/problem/saveChangeButton/SaveChangeButton.js
@@ -1,11 +1,0 @@
-import {EditorButton} from "../../commons/ProblemEditorPage/ProblemEditorPage";
-
-function SaveChangeButton () {
-    return (
-        <EditorButton
-            text={"Save Change"} buttonColor={"#96D745"} fontColor={"#FFFFFF"}
-        />
-    )
-}
-
-export default SaveChangeButton;
